+++ fmt/w64.c
@@ -68,7 +68,7 @@ static int w64_chunk_peek(struct w64_chunk *chunk, slurp_t *fp)
 	if (pos < 0)
 		return 0;
 
-	return ((uint64_t)pos <= (uint64_t)slurp_length(fp));
+	return ((uint64_t)pos <= slurp_length(fp));
 }
 
 static int w64_chunk_empty(struct w64_chunk *chunk)
